refactor: rename dir
