+++ packages/alert/__tests__/Alert.spec.tsx
@@ -3,7 +3,7 @@ import type { AlertProps } from '../src';
 import { axe } from 'jest-axe';
 import { it, expect, describe, vi } from 'vitest';
 
-import { render, screen, userEvent } from '../../../tests/utils';
+import { render, screen, userEvent } from '../../../test/utils';
 import { Alert, AlertKind, AlertSize } from '../src';
 
 const createComponent = (props?: AlertProps) => <Alert {...props}>Alert text!</Alert>;
