+++ packages/banner/__tests__/Banner.spec.tsx
@@ -2,7 +2,7 @@ import { AlertKind } from '@launchpad-ui/alert';
 import { axe } from 'jest-axe';
 import { it, expect, describe } from 'vitest';
 
-import { render, screen } from '../../../tests/utils';
+import { render, screen } from '../../../test/utils';
 import { Banner } from '../src';
 
 describe('Banner', () => {
