+++ packages/button/__tests__/Button.spec.tsx
@@ -2,7 +2,7 @@ import { Add } from '@launchpad-ui/icons';
 import { axe } from 'jest-axe';
 import { it, expect, describe, vi } from 'vitest';
 
-import { render, screen, userEvent } from '../../../tests/utils';
+import { render, screen, userEvent } from '../../../test/utils';
 import { Button, ButtonKind, ButtonSize, ButtonType } from '../src';
 
 describe('Button', () => {
