+++ packages/form/__tests__/IconField.spec.tsx
@@ -2,7 +2,7 @@ import { Info } from '@launchpad-ui/icons';
 import { axe } from 'jest-axe';
 import { it, expect, describe } from 'vitest';
 
-import { render, screen } from '../../../tests/utils';
+import { render, screen } from '../../../test/utils';
 import { IconField } from '../src';
 
 describe('IconField', () => {
