+++ packages/icons/__tests__/Icon.spec.tsx
@@ -1,7 +1,7 @@
 import { axe } from 'jest-axe';
 import { it, expect, describe } from 'vitest';
 
-import { render, screen } from '../../../tests/utils';
+import { render, screen } from '../../../test/utils';
 import { Add, IconSize, Info, KindIcon } from '../src';
 
 describe('Icon', () => {
