+++ packages/lozenge/__tests__/Lozenge.spec.tsx
@@ -1,7 +1,7 @@
 import { axe } from 'jest-axe';
 import { it, expect, describe, vi } from 'vitest';
 
-import { render, screen, userEvent } from '../../../tests/utils';
+import { render, screen, userEvent } from '../../../test/utils';
 import { Lozenge, LozengeKind, LozengeSize } from '../src';
 
 describe('Lozenge', () => {
