+++ packages/menu/__tests__/Menu.spec.tsx
@@ -4,7 +4,7 @@ import { Popover } from '@launchpad-ui/popover';
 import { axe } from 'jest-axe';
 import { it, expect, describe, vi } from 'vitest';
 
-import { render, screen, userEvent, waitFor } from '../../../tests/utils';
+import { render, screen, userEvent, waitFor } from '../../../test/utils';
 import { Menu, MenuDivider, MenuItem, MenuItemLink, MenuSearch } from '../src';
 
 type TestMenu = {
