+++ packages/notification/__tests__/Notification.spec.tsx
@@ -1,7 +1,7 @@
 import { axe } from 'jest-axe';
 import { it, expect, describe, vi } from 'vitest';
 
-import { render, screen, userEvent, waitFor } from '../../../tests/utils';
+import { render, screen, userEvent, waitFor } from '../../../test/utils';
 import { Notification, NotificationLevel } from '../src';
 
 const props = {
