+++ packages/progress/__tests__/Progress.spec.tsx
@@ -1,7 +1,7 @@
 import { axe } from 'jest-axe';
 import { it, expect, describe, vi } from 'vitest';
 
-import { render, screen, waitFor } from '../../../tests/utils';
+import { render, screen, waitFor } from '../../../test/utils';
 import { AnimationDelay, Progress, ProgressSize } from '../src';
 
 describe('Progress', () => {
