+++ packages/toggle/__tests__/Toggle.spec.tsx
@@ -1,6 +1,6 @@
 import { axe } from 'jest-axe';
 
-import { render, screen, userEvent, waitFor } from '../../../tests/utils';
+import { render, screen, userEvent, waitFor } from '../../../test/utils';
 import { Toggle } from '../src';
 
 describe('Toggle', () => {
